LPS-66928 Move DDM related data handler to the correct packages
